Slack's website preview always uses french
I pasted this in slack for coworkers, and it used french for the website description preview. Slack might just be dumb and use the first `<meta property="og:description">` it finds. Swapping language preferences in Slack did not seem to matter. In this case, you might want to put english first.