Add in webpack to build pipeline. Enable hot reload.
With this, webpack is integrated into the heroku build pipeline for both
CSS and Javascript. This allows transpiling the JS down into something
more compatible with older browsers like IE.

As a bonus, the webpack-dev-middleware has been added into the express
server for development mode with hot reload enabled. This means with
"npm run dev", developers will get 2 features:

  (a) the webpage with auto-reload when client JS is changed
  (b) style sheet changes are deployed without reload.
  (c) syntax errors in JS or css are displayed in an overlay in the
      browser.

It's possible to configure it so that the page does not reload when the
JS is changed, but given that there is no tie-in to rerender the
map list on reload, it doesn't help.