Shift out widget creation and use template strings.
@saikat and @nick were noting that non-devs have a hard time making
even small content fixes because of the dynamically generated DOM.

They suggested using handlebars in the client but looking at our usage
it seems a template string might sufficient. Prototyping by moving
the widget content generation out into a separate file and replacing
the explicity dom node construction with just an HTML blob that is
given the the browser parser.