Add cloud function for executing email link commands.
This set of functions takes an encrypted json blob, and on
successfuly decryption sends a signed version of it to the public
AppScript endpoint.  The AppScript is expected to verify the
signature and timestamp and on success, execute the command.

Detail errors are not reported back from the function to prevent
implementation details from leaking.